bittrex 분리
